Start with a redesign of the entire site
I did away with the leaves header on every page, but the leaves are now
featured much more prominently on the home page. The remainder of the
pages just contains a nicely coloured header.

Also, some pages now have TOCs with Bootstrap scrollspies (although the
contents of the TOC still need to be generated automatically - at the
moment they are hard-coded).

At the moment I only changed the English locale, so the other language
versions are terribly broken. The English one is also broken in quite
some places, but not terribly so.